+++ src/constants.js
@@ -8,6 +8,7 @@ export * from "./constants/misc.js";
 export * from "./constants/pets.js";
 export * from "./constants/pet-stats.js";
 export * from "./constants/random-emoji.js";
+export * from "./constants/skins.js";
 export * from "./constants/stats.js";
 export * from "./constants/tags.js";
 export * from "./constants/talismans.js";
