+++ views/stats.ejs
@@ -1118,14 +1118,10 @@ const getRarityUpgradeClass = item => {
             }
 
             let totalSkins = {}
-            for (const [petType, petSkins] of Object.entries(constants.pet_skins)) {
-              for (const [skin, skinData] of Object.entries(petSkins)) {
-                if (skinData.release < Date.now()) {
-                  totalSkins[`${petType}/${skin}`] =
-                    calculated.pets.find(
-                      (pet) => pet.type === petType && pet.skin === skin
-                    ) != undefined;
-                }
+            let badSkins = {}
+            for (const [skin, skinData] of Object.entries(constants.pet_skins)) {
+              if (skinData.release < Date.now()) {
+                totalSkins[skin] = calculated.pets.find((pet) => `PET_SKIN_${pet.skin}` === skin) != undefined;
               }
             }
             let userUniqueSkins = Object.values(totalSkins).filter(skin => skin).length
@@ -2330,7 +2326,8 @@ const getRarityUpgradeClass = item => {
     <%
       const clientConstants = {
         special_enchants: constants.special_enchants,
-        max_favorites: constants.max_favorites
+        max_favorites: constants.max_favorites,
+        pet_skins: constants.pet_skins,
       }
     %>
     const constants = JSON.parse(`<%- JSON.stringify(clientConstants).replace(/\\/g, '\\\\') %>`);
